# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "TowTruckMove REST API",
    "description": ""
  },
...
  "host": "localhost:8000",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/api/auth/verify/{token}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Entity"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}